Move Meta and LinksFunction to root
